fix(mdwt): correct heading_level type to i64 and fix cursor position after OTID inject
- Change heading_level from i32 to i64 to match sql schema
- Adjust cursor selection after OTID injection to prevent misalignment
- Update chin-tools submodule with I32 variant fix

+++ lib/backend/src/krate/mdwt/db.rs
@@ -593,7 +593,7 @@ impl MdwtMapper for KDb {
         struct ChildBlock {
             otid: TID,
             korder: i64,
-            heading_level: i32,
+            heading_level: i64,
         }
         let korder_name = cto.korder().field_name();
         let hl_name = cto.heading_level().field_name();
